XNR Resonance/Dissonance [Fixed]
There were a few glaring flaws and issues I had with the original file. In this version, neighboring nodes are found correctly, and if two neighbors have the same energy, then instead of randomly picking one of them, it gives energy to both. All that is needed at this point is to actually implement the proof of concept for inter-dimensional travel.